[WIP] Hacking on API things

+++ cloud/digitalocean/resources/ssh.go
@@ -60,10 +60,11 @@ func (r *SSH) Actual(known *cluster.Cluster) (cloud.Resource, error) {
 		strid := strconv.Itoa(ssh.ID)
 		actual.Name = ssh.Name
 		actual.CloudID = strid
-		actual.PublicKeyFingerprint = ssh.Fingerprint
-		actual.PublicKeyPath = known.SSH.PublicKeyPath
 		actual.PublicKeyData = ssh.PublicKey
+		actual.PublicKeyFingerprint = ssh.Fingerprint
 	}
+	actual.PublicKeyPath = known.SSH.PublicKeyPath
+	actual.User = known.SSH.User
 	r.CachedActual = actual
 	return actual, nil
 }
@@ -112,8 +113,11 @@ func (r *SSH) Apply(actual, expected cloud.Resource, applyCluster *cluster.Clust
 		if err != nil {
 			return nil, err
 		}
+		logger.Info("Using existing SSH Key [%s]", actual.(*SSH).Name)
+	}else {
+		logger.Info("Created SSH Key [%d]", key.ID)
 	}
-	logger.Info("Created SSH Key [%d]", key.ID)
+
 	id := strconv.Itoa(key.ID)
 	newResource := &SSH{
 		Shared: Shared{
@@ -127,24 +131,32 @@ func (r *SSH) Apply(actual, expected cloud.Resource, applyCluster *cluster.Clust
 	}
 	return newResource, nil
 }
-func (r *SSH) Delete(actual cloud.Resource, known *cluster.Cluster) error {
+func (r *SSH) Delete(actual cloud.Resource, known *cluster.Cluster) (cloud.Resource, error) {
 	logger.Debug("ssh.Delete")
-	deleteResource := actual.(*SSH)
-	if deleteResource.CloudID == "" {
-		return fmt.Errorf("Unable to delete ssh resource without Id [%s]", deleteResource.Name)
-	}
-	id, err := strconv.Atoi(known.SSH.Identifier)
-	if err != nil {
-		return err
-	}
+	force := false
+	if force {
+		deleteResource := actual.(*SSH)
+		if deleteResource.CloudID == "" {
+			return nil, fmt.Errorf("Unable to delete ssh resource without Id [%s]", deleteResource.Name)
+		}
+		id, err := strconv.Atoi(known.SSH.Identifier)
+		if err != nil {
+			return nil, err
+		}
 
-	_, err = Sdk.Client.Keys.DeleteByID(context.TODO(), id)
-	if err != nil {
-		return err
-	}
+		_, err = Sdk.Client.Keys.DeleteByID(context.TODO(), id)
+		if err != nil {
+			return nil, err
+		}
 
-	logger.Info("Deleted SSH Key [%d]", id)
-	return nil
+		logger.Info("Deleted SSH Key [%d]", id)
+	}
+	newResource := &SSH{}
+	newResource.Name = actual.(*SSH).Name
+	newResource.Tags = actual.(*SSH).Tags
+	newResource.User = actual.(*SSH).User
+	newResource.PublicKeyPath = actual.(*SSH).PublicKeyPath
+	return newResource, nil
 }
 
 func (r *SSH) Render(renderResource cloud.Resource, renderCluster *cluster.Cluster) (*cluster.Cluster, error) {
